agents/mcp-agents/src/text_mcp_agent/requirements.txt
# Text MCP Agent Test Dependencies

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# YAML processing
PyYAML>=6.0

# SAM dependencies (inherited)
google-adk-tools
solace-ai-connector

# Additional utilities
python-dotenv>=1.0.0
typing-extensions>=4.0.0
